Use servicePort in Ingress
Hi Mark,

I've made a small change to the Ingress template. The servicePort should be used in stead of the containerPort. We haven't discovered this issue yet because 99% of the applications are using the default 8080 port for both the service and the container. Now that we have applications with different ports, it is becoming a problem.

If you need more info, please let me know!

Kind regards,
Wesley